move event-box component to own file
